refactor: replace store
