+++ src/test/components/Pagination.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect } from 'vitest';
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { Pagination } from '../../components/Pagination/Pagination';
-import { store } from '../../../store';
+import { store } from '../../store/store';
 import { Provider } from 'react-redux';
 
 describe('Pagination', () => {
