+++ src/test/components/SearchBar.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect } from 'vitest';
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { SearchBar } from '../../components/SearchBar/SearchBar';
-import { store } from '../../../store';
+import { store } from '../../store/store';
 import { Provider } from 'react-redux';
 
 describe('SearchBar', () => {
