+++ src/test/pages/App.test.tsx
@@ -5,7 +5,7 @@ import userEvent from '@testing-library/user-event';
 import App from '../../App';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { store } from '../../../store';
+import { store } from '../../store/store';
 
 describe('App', () => {
   it('should be shown', async () => {
